HOLO-149 <refactor: 카테고리 검색 위치 기준 변경, 현재 위치에서 검색 버튼 삭제>
