--- src/app/map/_types/coordinates.ts
@@ -1,4 +1,0 @@
-export type Coordinates = {
-  lat: number;
-  lng: number;
-};
